Renamed play.js, moved host.js to a subfolder, Routing works, Animations don't
I added a routingWrapper file to the api folder that, when wrapped around a class component when the class is exported, allows routing to work through useRouter(). However, gsap animations are now broken, and I can't figure out how to route it so that host.js overlays on top of menu.js like it used to. Moving host.js into a menu subfolder didn't do anything.