Fixing the 1mb video limit: a story
The problem with large videos was that Lambda@Edge wouldn't allow a request body size larger than 1mb. So why not make the request to upload a video client-side instead of through our API? With this commit, when one makes a "submit" API request, they get a signed URL to upload a single video to S3. The client's browser then makes a PUT request locally to S3 as opposed to through the Next.js API, bypassing Lambda@Edge completely.